feat(tests): test SCENARIO.md
This patch adds a simple pytest suite in the new dir 'tests'.
It uses the handy
["CliRunner"](https://click.palletsprojects.com/en/7.x/testing/) in
the click.testing package to run each command as it would be run
on the command line.
The tests simply assert that each command succeeds.
Added a tox.ini, so you can use 'tox' (pip install tox) to
build and run in fresh virtual environment in one command (the
command is ... wait for it ... tox).